+++ R/helper.R
@@ -1,0 +1,6 @@
+recode = function(tab, levels) {
+  for (col in intersect(names(tab), names(levels))) {
+    set(tab, i = NULL, j = col, value = factor(tab[[col]], levels = levels[[col]]))
+  }
+  tab[]
+}