+++ man-roxygen/field_connector.R
@@ -1,0 +1,2 @@
+#' @field connector (`function()`)\cr
+#'   Function which is called to re-connect in case the connection became invalid.