+++ tests/testthat/test_as_duckdb_backend.R
@@ -1,33 +1,32 @@
-context("as_sqlite_backend")
+context("as_duckdb_backend")
 
-skip_if_not_installed("dplyr")
-skip_if_not_installed("dbplyr")
+skip_if_not_installed("duckdb")
 
 test_that("data", {
-  b = as_sqlite_backend(iris)
-  expect_is(b, "DataBackendDplyr")
+  b = as_duckdb_backend(iris)
+  expect_is(b, "DataBackendDuckDB")
   expect_backend(b)
   expect_iris_backend(b)
   disconnect(b)
 })
 
 test_that("DataBackend", {
-  b = as_sqlite_backend(mlr3::mlr_tasks$get("iris")$backend)
-  expect_is(b, "DataBackendDplyr")
+  b = as_duckdb_backend(mlr3::mlr_tasks$get("iris")$backend)
+  expect_is(b, "DataBackendDuckDB")
   expect_iris_backend(b)
   expect_backend(b)
   disconnect(b)
 })
 
 test_that("Task", {
-  b = as_sqlite_backend(mlr3::mlr_tasks$get("iris"))$backend
-  expect_is(b, "DataBackendDplyr")
+  b = as_duckdb_backend(mlr3::mlr_tasks$get("iris"))$backend
+  expect_is(b, "DataBackendDuckDB")
   expect_backend(b)
   disconnect(b)
 })
 
 test_that("connector is automatically set", {
-  b = as_sqlite_backend(iris)
+  b = as_duckdb_backend(iris)
   expect_function(b$connector)
   expect_set_equal(ls(environment(b$connector), all.names = TRUE), "path")
   disconnect(b)