+++ tests/testthat/test_train_predict_dplyr.R
@@ -1,3 +1,6 @@
+skip_if_not_installed("dplyr")
+skip_if_not_installed("dbplyr")
+
 b = as_sqlite_backend(iris)
 task = mlr3::TaskClassif$new("iris_sqlite", b, "Species")
 learner = mlr3::mlr_learners$get("classif.featureless")
@@ -11,8 +14,10 @@ test_that("single step train + predict", {
   expect_character(learner$errors, len = 0L)
 })
 
-test_that("resample work", {
-  rr = mlr3::resample(task, learner, mlr3::rsmp("cv", folds = 3))
+test_that("resample works", {
+  with_future("multisession", {
+    rr = mlr3::resample(task, learner, mlr3::rsmp("cv", folds = 3))
+  })
   expect_resample_result(rr)
 })